requirements.txt
lxml==5.2.2
numpy==2.0.0
opencv-python==4.10.0.84
pillow==10.4.0
python-pptx==0.6.23
XlsxWriter==3.2.0
